@@ -18,19 +18,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.0",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/http": "^7.2.15",
-    "@angular/material": "^7.3.7",
-    "@angular/material-moment-adapter": "^8.0.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/platform-server": "^7.2.15",
-    "@angular/router": "~7.2.0",
+    "@angular/animations": "^8.1.2",
+    "@angular/cdk": "^8.1.1",
+    "@angular/common": "^8.1.2",
+    "@angular/compiler": "^8.1.2",
+    "@angular/core": "^8.1.2",
+    "@angular/forms": "^8.1.2",
+    "@angular/material": "^8.1.1",
+    "@angular/material-moment-adapter": "^8.1.1",
+    "@angular/platform-browser": "^8.1.2",
+    "@angular/platform-browser-dynamic": "^8.1.2",
+    "@angular/platform-server": "^8.1.2",
+    "@angular/router": "^8.1.2",
     "@ngx-translate/core": "^11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
     "@types/jspdf": "^1.3.0",
@@ -39,7 +38,7 @@
     "bootstrap-notify": "^3.1.3",
     "chartist": "^0.11.2",
     "classlist.js": "^1.1.20150312",
-    "core-js": "^2.5.4",
+    "core-js": "^3.1.4",
     "crypto-js": "^3.1.9-1",
     "dotenv": "^8.0.0",
     "echarts": "^4.2.1",
@@ -50,24 +49,24 @@
     "moment": "^2.24.0",
     "mustache": "^3.0.1",
     "ng-recaptcha": "^4.3.0",
-    "ngx-echarts": "^4.1.1",
+    "ngx-echarts": "^4.2.1",
     "ngx-toastr": "^10.0.4",
     "node-sass": "^4.12.0",
     "perfect-scrollbar": "^1.4.0",
     "popper.js": "^1.15.0",
-    "rxjs": "~6.3.3",
-    "rxjs-compat": "~6.3.3",
+    "rxjs": "^6.5.2",
+    "rxjs-compat": "^6.5.2",
     "sass-loader": "^7.1.0",
     "saturn-datepicker": "^7.4.0",
     "tslib": "^1.9.0",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "~0.8.26"
+    "web-animations-js": "^2.3.2",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.3",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
+    "@angular-devkit/build-angular": "~0.801.2",
+    "@angular/cli": "~8.1.2",
+    "@angular/compiler-cli": "~8.1.2",
+    "@angular/language-service": "~8.1.2",
     "@types/bootstrap": "^4.3.0",
     "@types/chartist": "^0.9.46",
     "@types/datatables.net": "^1.10.17",
@@ -76,7 +75,7 @@
     "@types/jasminewd2": "~2.0.3",
     "@types/jquery": "^3.3.29",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.0.0",
@@ -87,6 +86,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "typescript": "~3.4.5"
   }
 }
