@@ -1,10 +1,10 @@
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
 
-import {MatPaginatorIntl} from '@angular/material';
+import { MatPaginatorIntl } from '@angular/material';
 import { TranslateService } from '@ngx-translate/core';
 
-import {PaginatorIntlService} from './config-matpaginator';
+import { PaginatorIntlService } from './config-matpaginator';
 
 @NgModule({
     declarations: [],
@@ -12,15 +12,16 @@ import {PaginatorIntlService} from './config-matpaginator';
         CommonModule
     ],
     providers: [
+        PaginatorIntlService,
         {
             provide: MatPaginatorIntl,
             useFactory: (translate) => {
-              const service = new PaginatorIntlService();
-              service.injectTranslateService(translate);
-              return service;
+                const service = new PaginatorIntlService();
+                service.injectTranslateService(translate);
+                return service;
             },
             deps: [TranslateService]
-          }
+        }
     ]
 })
 export class ModulesModule {
