@@ -1,15 +1,16 @@
 import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: 'loading-component',
-  templateUrl: './loading-component.component.html',
-  styleUrls: ['./loading-component.component.scss']
+    selector: 'loading-component',
+    templateUrl: './loading-component.component.html',
+    styleUrls: ['./loading-component.component.scss']
 })
 export class LoadingComponentComponent implements OnInit {
 
-  constructor() { }
+    constructor() {
+    }
 
-  ngOnInit() {
-  }
+    ngOnInit() {
+    }
 
 }
