@@ -1,7 +1,8 @@
-import {Injectable} from '@angular/core';
-import {HttpClient} from "@angular/common/http";
-import {NotificationEmail} from "../../modules/evaluation/models/notification-email";
-import {environment} from "../../../environments/environment";
+import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+
+import { NotificationEmail } from '../../modules/evaluation/models/notification-email';
+import { environment } from '../../../environments/environment';
 
 @Injectable()
 export class NotificationService {
